+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/elements/ClickableElement.kt
@@ -37,8 +37,10 @@ fun ClickableElement(
             descriptionId = descriptionId
         )
         Icon(
-            modifier = Modifier.size(size = 42.dp).padding(16.dp),
-            painter = painterResource(DesignSystem.drawable.ic_navigate_icon),
+            modifier = Modifier
+                .size(size = 42.dp)
+                .padding(16.dp),
+            painter = painterResource(DesignSystem.drawable.ic_navigate),
             tint = LocalPallet.current.iconTint30,
             contentDescription = null
         )